update migrations for item update delete constraint update assign and buy item

+++ src/pages/api/admin/assign-item/remove.ts
@@ -2,7 +2,7 @@ import { getSession } from 'next-auth/react';
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { prisma } from 'db';
-import { forbiddenResponse, stillInUseResponse, successResponse } from 'utils/constant';
+import { forbiddenResponse, successResponse } from 'utils/constant';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
 	const { locId, itemId } = req.body;
@@ -12,17 +12,30 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
 	if (!session) return res.json(forbiddenResponse);
 
 	if (locId && itemId) {
-		// const employees = await prisma.location.findMany({ where: { titleId: id } });
-
-		// if (employees.length > 0) {
-		// 	return res.json({ ...stillInUseResponse });
-		// }
-
-		const update = await prisma.itemLocation.deleteMany({
-			where: { locId, itemId },
+		const detail = await prisma.itemLocation.findUnique({
+			where: {
+				locId_itemId: {
+					locId,
+					itemId,
+				},
+			},
 		});
 
-		return res.json({ ...successResponse, data: update });
+		if (!detail) {
+			return res.json({ code: 404, message: 'data not found' });
+		}
+
+		const process = await prisma.$transaction([
+			prisma.itemLocation.deleteMany({
+				where: { locId, itemId },
+			}),
+			prisma.item.updateMany({
+				where: { id: itemId },
+				data: { assignQty: { decrement: detail?.qty || 0 } },
+			}),
+		]);
+
+		return res.json({ ...successResponse, data: process });
 	}
 
 	return res.json({ code: 500, message: 'location id & item id is required' });
