+++ src/pages/api/admin/location/remove.ts
@@ -12,11 +12,11 @@ export default async function handle(req: NextApiRequest, res: NextApiResponse) 
 	if (!session) return res.json(forbiddenResponse);
 
 	if (id) {
-		// const employees = await prisma.location.findMany({ where: { titleId: id } });
+		const itemLocation = await prisma.itemLocation.findMany({ where: { locId: id } });
 
-		// if (employees.length > 0) {
-		// 	return res.json({ ...stillInUseResponse });
-		// }
+		if (itemLocation.length > 0) {
+			return res.json({ ...stillInUseResponse });
+		}
 
 		const update = await prisma.location.delete({
 			where: { id },
